moved config to types
